Rename log package to wal
